Update Google Breakpad from SVN r480 to r967
This fixes the build on OS X 10.7 with Clang, along with various bug fixes.

 * Update google-breakpad to upstream r967
 * Adjust CMake build script for Mac breakpad client
 * Adapt BreakpadCrashHandler to new ExceptionHandler constructor signature.

MD-18534 #time 1h
unit-tests: other: We need to manually verify that usable crash reports are generated on Windows from the new breakpad version.

+++ src/client/minidump_file_writer.h
@@ -151,7 +151,8 @@ class UntypedMDRVA {
 
   // Return size and position
   inline MDLocationDescriptor location() const {
-    MDLocationDescriptor location = { size_, position_ };
+    MDLocationDescriptor location = { static_cast<u_int32_t>(size_),
+                                      position_ };
     return location;
   }
 
@@ -218,7 +219,7 @@ class TypedMDRVA : public UntypedMDRVA {
   // Allocate an array of |count| elements of |size| after object of MDType
   // Must not call more than once.
   // Return true on success, or false on failure
-  bool AllocateObjectAndArray(unsigned int count, size_t size);
+  bool AllocateObjectAndArray(size_t count, size_t size);
 
   // Copy |item| to |index|
   // Must have been allocated using AllocateArray().
